Adicionado tratamento de exceção

+++ src/main/java/com/cursodevsuperior/dscommerce/services/CategoryService.java
@@ -1,14 +1,20 @@
 package com.cursodevsuperior.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.cursodevsuperior.dscommerce.dto.CategoryDTO;
 import com.cursodevsuperior.dscommerce.entities.Category;
 import com.cursodevsuperior.dscommerce.repositories.CategoryRepository;
+import com.cursodevsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.cursodevsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class CategoryService {
@@ -24,7 +30,8 @@ public class CategoryService {
 
 	@Transactional(readOnly = true)
 	public CategoryDTO findById(Long id) {
-		Category result = repository.findById(id).get();
+		Category result = repository.findById(id)
+						.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new CategoryDTO(result);
 	}
 
@@ -37,14 +44,27 @@ public class CategoryService {
 
 	@Transactional
 	public CategoryDTO update(Long id, CategoryDTO dto) {
-		Category entity = repository.getReferenceById(id);
-		copyDataAndSave(dto, entity);
-		return new CategoryDTO(entity);
+		try {
+			Category entity = repository.getReferenceById(id);
+			copyDataAndSave(dto, entity);
+			return new CategoryDTO(entity);
+		}
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Violação de integridade relacional");
+		}
 	}
 
 	private void copyDataAndSave(CategoryDTO dto, Category entity) {
