+++ src/main/java/com/cursodevsuperior/dscommerce/services/ProductService.java
@@ -1,9 +1,11 @@
 package com.cursodevsuperior.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.cursodevsuperior.dscommerce.dto.CategoryDTO;
@@ -12,6 +14,10 @@ import com.cursodevsuperior.dscommerce.dto.ProductMinDTO;
 import com.cursodevsuperior.dscommerce.entities.Category;
 import com.cursodevsuperior.dscommerce.entities.Product;
 import com.cursodevsuperior.dscommerce.repositories.ProductRepository;
+import com.cursodevsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.cursodevsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -27,7 +33,8 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product result = repository.findById(id).get();
+		Product result = repository.findById(id)
+						.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProductDTO(result);
 	}
 
@@ -40,14 +47,27 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDataAndSave(dto, entity);
-		return new ProductDTO(entity);
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDataAndSave(dto, entity);
+			return new ProductDTO(entity);
+		}
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Violação de integridade relacional");
+		}
 	}
 	
 	private void copyDataAndSave(ProductDTO dto, Product entity) {
